refactor: move getProperties
